Moved TC15 and TC16 to separate integration testing suite
