+++ docs/bva/MainMenuController.md
@@ -68,11 +68,3 @@
 - **TC14: startNewGame_WithOneMoreThanMaximumPlayers_ThrowsException** ( :white_check_mark: )
     - **State of the system**: player config count = 5
     - **Expected output**: throws `IllegalArgumentException`
-
-- **TC15: startNewGame_WithValidSetup_SetsGameStatusToInProgress** ( :white_check_mark: )
-    - **State of the system**: player config count = 2, all player configs are valid
-    - **Expected output**: game status becomes `GameStatus.IN_PROGRESS`
-
-- **TC16: startNewGame_WithValidSetup_SetsFirstPlayerAsCurrentPlayer** ( :white_check_mark: )
-    - **State of the system**: player config count = 2, all player configs are valid
-    - **Expected output**: current player is the first configured player
